+++ Backtrace.Tests/Events/ApiEventsTests.cs
@@ -51,12 +51,12 @@ namespace Backtrace.Tests.Events
             //mock database
             var database = new Mock<IBacktraceDatabase>();
             database.Setup(n =>
-                n.Add(It.IsAny<BacktraceReportBase>(),
+                n.Add(It.IsAny<BacktraceReport>(),
                     It.IsAny<Dictionary<string, object>>(),
                     It.IsAny<MiniDumpType>()));
 
             database.Setup(n =>
-               n.Delete(It.IsAny<BacktraceDatabaseEntry>()));
+               n.Delete(It.IsAny<BacktraceDatabaseRecord>()));
 
             //setup new client
             _backtraceClient = new BacktraceClient(credentials, database: database.Object, reportPerMin: 0)