+++ Backtrace/BacktraceDatabase.cs
@@ -28,7 +28,7 @@ namespace Backtrace
         public IBacktraceApi BacktraceApi { get; set; }
 
         /// <summary>
-        /// Database context - in memory cache and entry operations
+        /// Database context - in memory cache and record operations
         /// </summary>
         internal IBacktraceDatabaseContext BacktraceDatabaseContext { get; set; }
 
@@ -86,7 +86,7 @@ namespace Backtrace
             }
             DatabaseSettings = databaseSettings;
             BacktraceDatabaseContext = new BacktraceDatabaseContext(DatabasePath, DatabaseSettings.RetryLimit, DatabaseSettings.RetryOrder);
-            BacktraceDatabaseFileContext = new BacktraceDatabaseFileContext(DatabasePath);
+            BacktraceDatabaseFileContext = new BacktraceDatabaseFileContext(DatabasePath, DatabaseSettings.MaxDatabaseSize, DatabaseSettings.MaxRecordCount);
         }
 
         /// <summary>
@@ -158,12 +158,13 @@ namespace Backtrace
         public void Clear()
         {
             BacktraceDatabaseContext?.Clear();
+            BacktraceDatabaseFileContext?.Clear();
         }
 
         /// <summary>
         /// Add new report to BacktraceDatabase
         /// </summary>
-        public BacktraceDatabaseEntry Add(BacktraceReportBase backtraceReport, Dictionary<string, object> attributes, MiniDumpType miniDumpType = MiniDumpType.Normal)
+        public BacktraceDatabaseRecord Add(BacktraceReport backtraceReport, Dictionary<string, object> attributes, MiniDumpType miniDumpType = MiniDumpType.Normal)
         {
             if (!_enable)
             {
@@ -171,7 +172,11 @@ namespace Backtrace
             }
             if (BacktraceDatabaseContext.Count() + 1 > DatabaseSettings.MaxRecordCount && DatabaseSettings.MaxRecordCount != 0)
             {
-                throw new ArgumentException("Maximum number of entries available in BacktraceDatabase");
+                throw new ArgumentException("Maximum number of records available in BacktraceDatabase");
+            }
+            if (BacktraceDatabaseContext.GetSize() > DatabaseSettings.MaxDatabaseSize && DatabaseSettings.MaxDatabaseSize != 0)
+            {
+                throw new ArgumentException("You don't have enought space in database for more records");
             }
             if (miniDumpType != MiniDumpType.None)
             {
@@ -188,15 +193,19 @@ namespace Backtrace
 
 
         /// <summary>
-        /// Get all stored reports in BacktraceDatabase
+        /// Get all stored records in BacktraceDatabase
         /// </summary>
-        /// <returns>All stored reports in BacktraceDatabase</returns>
-        public IEnumerable<BacktraceDatabaseEntry> Get() => BacktraceDatabaseContext?.Get() ?? new List<BacktraceDatabaseEntry>();
-
-        public void Delete(BacktraceDatabaseEntry entry) => BacktraceDatabaseContext?.Delete(entry);
+        /// <returns>All stored records in BacktraceDatabase</returns>
+        public IEnumerable<BacktraceDatabaseRecord> Get() => BacktraceDatabaseContext?.Get() ?? new List<BacktraceDatabaseRecord>();
 
         /// <summary>
-        /// Send and delete all entries from database
+        /// Delete single record from database
+        /// </summary>
+        /// <param name="record">Record to delete</param>
+        public void Delete(BacktraceDatabaseRecord record) => BacktraceDatabaseContext?.Delete(record);
+
+        /// <summary>
+        /// Send and delete all records from database
         /// </summary>
         public void Flush()
         {
@@ -204,12 +213,12 @@ namespace Backtrace
             {
                 throw new ArgumentException("BacktraceApi is required if you want to use Flush method");
             }
-            var entry = BacktraceDatabaseContext?.FirstOrDefault();
-            while (entry != null)
+            var record = BacktraceDatabaseContext?.FirstOrDefault();
+            while (record != null)
             {
-                var backtraceData = entry.BacktraceData;
-                Delete(entry);
-                entry = BacktraceDatabaseContext.FirstOrDefault();
+                var backtraceData = record.BacktraceData;
+                Delete(record);
+                record = BacktraceDatabaseContext.FirstOrDefault();
                 if (backtraceData != null)
                 {
                     BacktraceApi.Send(backtraceData);
@@ -218,7 +227,7 @@ namespace Backtrace
         }
 #if !NET35
         /// <summary>
-        /// Send and asynchronous delete all entries from database
+        /// Send and asynchronous delete all records from database
         /// </summary>
         public async Task FlushAsync()
         {
@@ -226,12 +235,12 @@ namespace Backtrace
             {
                 throw new ArgumentException("BacktraceApi is required if you want to use Flush method");
             }
-            var entry = BacktraceDatabaseContext?.FirstOrDefault();
-            while (entry != null)
+            var record = BacktraceDatabaseContext?.FirstOrDefault();
+            while (record != null)
             {
-                var backtraceData = entry.BacktraceData;
-                Delete(entry);
-                entry = BacktraceDatabaseContext.FirstOrDefault();
+                var backtraceData = record.BacktraceData;
+                Delete(record);
+                record = BacktraceDatabaseContext.FirstOrDefault();
                 if (backtraceData != null)
                 {
                     await BacktraceApi.SendAsync(backtraceData);
@@ -244,33 +253,33 @@ namespace Backtrace
             if (!BacktraceDatabaseContext.Any() || _timerBackgroundWork) return;
             _timerBackgroundWork = true;
             _timer.Stop();
-            //read first entry (keep in mind LIFO and FIFO settings) from memory database
-            var entry = BacktraceDatabaseContext.FirstOrDefault();
-            while (entry != null)
+            //read first record (keep in mind LIFO and FIFO settings) from memory database
+            var record = BacktraceDatabaseContext.FirstOrDefault();
+            while (record != null)
             {
-                var backtraceData = entry.BacktraceData;
+                var backtraceData = record.BacktraceData;
                 //meanwhile someone delete data from a disk
                 if (backtraceData == null || backtraceData.Report == null)
                 {
-                    Delete(entry);
+                    Delete(record);
                 }
                 else
                 {
-                    //send entry from database to API
+                    //send record from database to API
                     var result = await BacktraceApi.SendAsync(backtraceData);
                     if (result.Status == BacktraceResultStatus.Ok)
                     {
-                        Delete(entry);
+                        Delete(record);
                     }
                     else
                     {
-                        entry.Dispose();
+                        record.Dispose();
                         BacktraceDatabaseContext.IncrementBatchRetry();
                         break;
                     }
 
                 }
-                entry = BacktraceDatabaseContext.FirstOrDefault();
+                record = BacktraceDatabaseContext.FirstOrDefault();
             }
             _timer.Start();
             _timerBackgroundWork = false;
@@ -281,33 +290,33 @@ namespace Backtrace
             if (!BacktraceDatabaseContext.Any() || _timerBackgroundWork) return;
             _timerBackgroundWork = true;
             _timer.Stop();
-            //read first entry (keep in mind LIFO and FIFO settings) from memory database
-            var entry = BacktraceDatabaseContext.FirstOrDefault();
-            while (entry != null)
+            //read first record (keep in mind LIFO and FIFO settings) from memory database
+            var record = BacktraceDatabaseContext.FirstOrDefault();
+            while (record != null)
             {
-                var backtraceData = entry.BacktraceData;
+                var backtraceData = record.BacktraceData;
                 //meanwhile someone delete data from a disk
                 if (backtraceData == null || backtraceData.Report == null)
                 {
-                    Delete(entry);
+                    Delete(record);
                 }
                 else
                 {
-                    //send entry from database to API
+                    //send record from database to API
                     var result = BacktraceApi.Send(backtraceData);
                     if (result.Status == BacktraceResultStatus.Ok)
                     {
-                        Delete(entry);
+                        Delete(record);
                     }
                     else
                     {
-                        entry.Dispose();
+                        record.Dispose();
                         BacktraceDatabaseContext.IncrementBatchRetry();
                         break;
 
                     }
                 }
-                entry = BacktraceDatabaseContext.FirstOrDefault();
+                record = BacktraceDatabaseContext.FirstOrDefault();
             }
             _timerBackgroundWork = false;
             _timer.Start();
@@ -320,7 +329,7 @@ namespace Backtrace
         /// <param name="backtraceReport">Current report</param>
         /// <param name="miniDumpType">Generated minidump type</param>
         /// <returns>Path to minidump file</returns>
-        private string GenerateMiniDump(BacktraceReportBase backtraceReport, MiniDumpType miniDumpType)
+        private string GenerateMiniDump(BacktraceReport backtraceReport, MiniDumpType miniDumpType)
         {
             //note that every minidump file generated by app ends with .dmp extension
             //its important information if you want to clear minidump file
@@ -340,38 +349,60 @@ namespace Backtrace
         }
 
         /// <summary>
-        /// Get total number of entries in database
+        /// Get total number of records in database
         /// </summary>
-        /// <returns>Total number of entries</returns>
+        /// <returns>Total number of records</returns>
         internal int Count() => BacktraceDatabaseContext.Count();
 
         /// <summary>
-        /// Detect all orphaned minidump files
+        /// Detect all orphaned minidump and files
         /// </summary>
         private void RemoveOrphaned()
         {
-            var entries = BacktraceDatabaseContext.Get();
-            BacktraceDatabaseFileContext.RemoveOrphaned(entries);
+            var records = BacktraceDatabaseContext.Get();
+            BacktraceDatabaseFileContext.RemoveOrphaned(records);
         }
 
         /// <summary>
-        /// Load all reports stored in path passed by user
+        /// Load all records stored in database path
         /// </summary>
         private void LoadReports()
         {
-            var files = BacktraceDatabaseFileContext.GetEntries();
+            var files = BacktraceDatabaseFileContext.GetRecords();
             foreach (var file in files)
             {
-                var entry = BacktraceDatabaseEntry.ReadFromFile(file);
-                if (!entry.Valid())
+                var record = BacktraceDatabaseRecord.ReadFromFile(file);
+                if (!record.Valid())
                 {
-                    entry.Delete();
+                    record.Delete();
                     continue;
                 }
-                BacktraceDatabaseContext.Add(entry);
-                entry.Dispose();
+                BacktraceDatabaseContext.Add(record);
+                if (!ValidDatabaseSize())
+                {
+                    throw new ArgumentException("Database directory has too many records or database size is bigger than in option declaration.");
+                }
+                record.Dispose();
             }
         }
+        /// <summary>
+        /// Check current size of database
+        /// </summary>
+        /// <returns>False if BacktraceDatabase doesn't have more free space</returns>
+        private bool ValidDatabaseSize()
+        {
+            return ((BacktraceDatabaseContext.GetSize() < DatabaseSettings.MaxDatabaseSize || DatabaseSettings.MaxDatabaseSize == 0)
+                || (BacktraceDatabaseContext.GetTotalNumberOfRecords() < DatabaseSettings.MaxRecordCount || DatabaseSettings.MaxDatabaseSize == 0));
+        }
+
+        /// <summary>
+        /// Valid database consistency requirements
+        /// </summary>
+        public bool ValidConsistency()
+        {
+            return BacktraceDatabaseFileContext.ValidFileConsistency();
+        }
+
 
         #region dispose
         private bool _disposed = false; // To detect redundant calls
@@ -398,6 +429,15 @@ namespace Backtrace
                 }
                 _disposed = true;
             }
+        }
+
+        /// <summary>
+        /// Get database size
+        /// </summary>
+        /// <returns></returns>
+        public long GetDatabaseSize()
+        {
+            return BacktraceDatabaseContext.GetSize();
         }
 
         ~BacktraceDatabase()