+++ Backtrace/Interfaces/Database/IBacktraceDatabaseRecordWriter.cs
@@ -5,12 +5,12 @@ using System.Text;
 [assembly: System.Runtime.CompilerServices.InternalsVisibleTo("DynamicProxyGenAssembly2")]
 namespace Backtrace.Interfaces.Database
 {
-    internal interface IBacktraceDatabaseEntryWriter
+    internal interface IBacktraceDatabaseRecordWriter
     {
         string Write(object data, string prefix);
         string Write(byte[] data, string prefix);
         void SaveTemporaryFile(string path, byte[] file);
         string ToJsonFile(object data);
-        void SaveValidReport(string sourcePath, string destinationPath);
+        void SaveValidRecord(string sourcePath, string destinationPath);
     }
 }