+++ Backtrace/Interfaces/IBacktraceDatabaseContext.cs
@@ -11,55 +11,55 @@ namespace Backtrace.Interfaces
     internal interface IBacktraceDatabaseContext : IDisposable
     {
         /// <summary>
-        /// Add new report to Database
+        /// Add new record to Database
         /// </summary>
         /// <param name="backtraceData">Diagnostic data</param>
-        BacktraceDatabaseEntry Add(BacktraceData backtraceData);
+        BacktraceDatabaseRecord Add(BacktraceData backtraceData);
 
         /// <summary>
         /// Add new data to database
         /// </summary>
-        /// <param name="backtraceDatabaseEntry">Database entry</param>
-        BacktraceDatabaseEntry Add(BacktraceDatabaseEntry backtraceDatabaseEntry);
+        /// <param name="backtraceDatabaseRecord">Database record</param>
+        BacktraceDatabaseRecord Add(BacktraceDatabaseRecord backtraceDatabaseRecord);
 
         /// <summary>
-        /// Get first entry or null
+        /// Get first record or null
         /// </summary>
-        /// <returns>First existing entry in database store</returns>
-        BacktraceDatabaseEntry FirstOrDefault();
+        /// <returns>First existing record in database store</returns>
+        BacktraceDatabaseRecord FirstOrDefault();
 
         /// <summary>
-        /// Get last entry or null
+        /// Get last record or null
         /// </summary>
-        /// <returns>Last existing entry in database store</returns>
-        BacktraceDatabaseEntry LastOrDefault();
+        /// <returns>Last existing record in database store</returns>
+        BacktraceDatabaseRecord LastOrDefault();
 
         /// <summary>
-        /// Get all repots stored in Database
+        /// Get all records stored in Database
         /// </summary>
-        IEnumerable<BacktraceDatabaseEntry> Get();
+        IEnumerable<BacktraceDatabaseRecord> Get();
 
         /// <summary>
-        /// Delete database entry by using BacktraceDatabaseEntry
+        /// Delete database record by using BacktraceDatabaseRecord
         /// </summary>
-        /// <param name="entry">Database entry</param>
-        void Delete(BacktraceDatabaseEntry entry);
+        /// <param name="record">Database record</param>
+        void Delete(BacktraceDatabaseRecord record);
 
         /// <summary>
-        /// Check if any similar entry exists
+        /// Check if any similar record exists
         /// </summary>
-        /// <param name="n">Compared entry</param>
-        bool Any(BacktraceDatabaseEntry n);
+        /// <param name="n">Compared record</param>
+        bool Any(BacktraceDatabaseRecord n);
 
         /// <summary>
-        /// Check if any similar entry exists
+        /// Check if any similar record exists
         /// </summary>
         bool Any();
 
         /// <summary>
-        /// Get total count of entries
+        /// Get total count of records
         /// </summary>
-        /// <returns>Total number of entries</returns>
+        /// <returns>Total number of records</returns>
         int Count();
 
         /// <summary>
@@ -68,10 +68,21 @@ namespace Backtrace.Interfaces
         void Clear();
 
         /// <summary>
-        /// Increment retry time for all entries
+        /// Increment record time for all records
         /// </summary>
         void IncrementBatchRetry();
 
+        /// <summary>
+        /// Get database size
+        /// </summary>
+        /// <returns>Database size</returns>
+        long GetSize();
+
+        /// <summary>
+        /// Get total number of records stored in database
+        /// </summary>
+        /// <returns>Total number of records</returns>
+        int GetTotalNumberOfRecords();
 
         /// <summary>
         /// Get all repots stored in Database