+++ Backtrace/Model/Database/BacktraceDatabaseRecord.cs
@@ -12,27 +12,27 @@ using System.Text;
 namespace Backtrace.Model.Database
 {
     /// <summary>
-    /// Single entry in BacktraceDatabase
+    /// Single record in BacktraceDatabase
     /// </summary>
-    public class BacktraceDatabaseEntry: IDisposable
+    public class BacktraceDatabaseRecord : IDisposable
     {
         /// <summary>
-        /// Entry Id
+        /// Id
         /// </summary>
         [JsonProperty]
         public Guid Id { get; set; } = Guid.NewGuid();
 
         /// <summary>
-        /// Check if current entry is in use
+        /// Check if current record is in use
         /// </summary>
         [JsonIgnore]
         internal bool Locked { get; set; } = false;
 
         /// <summary>
-        /// Path to json stored all information about current entry
+        /// Path to json stored all information about current record
         /// </summary>
-        [JsonProperty(PropertyName = "entryName")]
-        internal string EntryPath { get; set; }
+        [JsonProperty(PropertyName = "recordName")]
+        internal string RecordPath { get; set; }
 
         /// <summary>
         /// Path to a diagnostic data json
@@ -53,10 +53,16 @@ namespace Backtrace.Model.Database
         internal string ReportPath { get; set; }
 
         /// <summary>
-        /// Stored entry
+        /// Total size of record
+        /// </summary>
+        [JsonProperty(PropertyName = "size")]
+        internal long Size { get; set; }
+
+        /// <summary>
+        /// Stored record
         /// </summary>
         [JsonIgnore]
-        internal virtual BacktraceData Entry { get; set; }
+        internal virtual BacktraceData Record { get; set; }
 
         /// <summary>
         /// Path to database directory
@@ -65,22 +71,22 @@ namespace Backtrace.Model.Database
         private readonly string _path = string.Empty;
 
         /// <summary>
-        /// Entry writer
+        /// Record writer
         /// </summary>
         [JsonIgnore]
-        internal IBacktraceDatabaseEntryWriter EntryWriter;
+        internal IBacktraceDatabaseRecordWriter RecordWriter;
 
         /// <summary>
-        /// Get valid BacktraceData from current entry
+        /// Get valid BacktraceData from current record
         /// </summary>
         [JsonIgnore]
         public virtual BacktraceData BacktraceData
         {
             get
             {
-                if (Entry != null)
+                if (Record != null)
                 {
-                    return Entry;
+                    return Record;
                 }
                 if (!Valid())
                 {
@@ -99,7 +105,7 @@ namespace Backtrace.Model.Database
                     try
                     {
                         var diagnosticData = JsonConvert.DeserializeObject<BacktraceData>(diagnosticDataJson);
-                        var report = JsonConvert.DeserializeObject<BacktraceReportBase>(reportJson);
+                        var report = JsonConvert.DeserializeObject<BacktraceReport>(reportJson);
                         //add report to diagnostic data
                         //we don't store report with diagnostic data in the same json
                         //because we have easier way to serialize and deserialize data
@@ -119,22 +125,22 @@ namespace Backtrace.Model.Database
         /// Constructor for serialization purpose
         /// </summary>
         [JsonConstructor]
-        internal BacktraceDatabaseEntry()
+        internal BacktraceDatabaseRecord()
         {
-            EntryPath = $"{Id}-entry.json";
+            RecordPath = $"{Id}-record.json";
         }
 
         /// <summary>
-        /// Create new instance of database entry
+        /// Create new instance of database record
         /// </summary>
         /// <param name="data">Diagnostic data</param>
         /// <param name="path">database path</param>
-        public BacktraceDatabaseEntry(BacktraceData data, string path)
+        public BacktraceDatabaseRecord(BacktraceData data, string path)
         {
             Id = data.Uuid;
-            Entry = data;
+            Record = data;
             _path = path;
-            EntryWriter = new BacktraceDatabaseEntryWriter(path);
+            RecordWriter = new BacktraceDatabaseRecordWriter(path);
         }
 
         /// <summary>
@@ -145,11 +151,22 @@ namespace Backtrace.Model.Database
         {
             try
             {
-                DiagnosticDataPath = EntryWriter.Write(Entry, $"{Id}-attachment");
-                ReportPath = EntryWriter.Write(Entry.Report, $"{Id}-report");
-                MiniDumpPath = Entry.Report?.MinidumpFile ?? string.Empty;
-                EntryPath = Path.Combine(_path, $"{Id}-entry.json");
-                EntryWriter.Write(this, $"{Id}-entry");
+                DiagnosticDataPath = Save(Record, $"{Id}-attachment");
+                ReportPath = Save(Record.Report, $"{Id}-report");
+                
+                // get minidump information
+                MiniDumpPath = Record.Report?.MinidumpFile ?? string.Empty;
+                Size += MiniDumpPath == string.Empty ? 0 : new FileInfo(MiniDumpPath).Length;
+
+                //save record
+                RecordPath = Path.Combine(_path, $"{Id}-record.json");
+                //check current record size
+                var json = JsonConvert.SerializeObject(this);
+                byte[] file = Encoding.UTF8.GetBytes(json);
+                //add record size
+                Size += file.Length;
+                //save it again with actual record size
+                RecordWriter.Write(this, $"{Id}-record");
                 return true;
             }
             catch (IOException)
@@ -165,27 +182,45 @@ namespace Backtrace.Model.Database
         }
 
         /// <summary>
-        /// Check if all necessary files declared on entry exists
+        /// Save single file from database record
         /// </summary>
-        /// <returns>True if entry is valid</returns>
+        /// <param name="data">single file (json/dmp)</param>
+        /// <param name="prefix">file prefix</param>
+        /// <returns>path to file</returns>
+        private string Save(object data, string prefix)
+        {
+            if (data == null)
+            {
+                return string.Empty;
+            }
+            var json = JsonConvert.SerializeObject(data);
+            byte[] file = Encoding.UTF8.GetBytes(json);
+            Size += file.Length;
+            return RecordWriter.Write(file, prefix);
+        }
+
+        /// <summary>
+        /// Check if all necessary files declared on record exists
+        /// </summary>
+        /// <returns>True if record is valid</returns>
         public bool Valid()
         {
             return File.Exists(DiagnosticDataPath) && File.Exists(ReportPath);
         }
 
         /// <summary>
-        /// Delete all entry files
+        /// Delete all record files
         /// </summary>
         internal virtual void Delete()
         {
             Delete(MiniDumpPath);
             Delete(ReportPath);
             Delete(DiagnosticDataPath);
-            Delete(EntryPath);
+            Delete(RecordPath);
         }
 
         /// <summary>
-        /// Delete single file on database entry
+        /// Delete single file on database record
         /// </summary>
         /// <param name="path">path to file</param>
         private void Delete(string path)
@@ -208,18 +243,18 @@ namespace Backtrace.Model.Database
         }
 
         /// <summary>
-        /// Read single entry from file
+        /// Read single record from file
         /// </summary>
         /// <param name="file">Current file</param>
-        /// <returns>Saved database entry</returns>
-        internal static BacktraceDatabaseEntry ReadFromFile(FileInfo file)
+        /// <returns>Saved database record</returns>
+        internal static BacktraceDatabaseRecord ReadFromFile(FileInfo file)
         {
             using (StreamReader streamReader = file.OpenText())
             {
                 var json = streamReader.ReadToEnd();
                 try
                 {
-                    return JsonConvert.DeserializeObject<BacktraceDatabaseEntry>(json);
+                    return JsonConvert.DeserializeObject<BacktraceDatabaseRecord>(json);
                 }
                 catch (SerializationException)
                 {
@@ -240,7 +275,7 @@ namespace Backtrace.Model.Database
             if (disposing)
             {
                 Locked = false;
-                Entry = null;
+                Record = null;
             }
         }
         #endregion